(PC-35683)[PRO] feat: Rewrites <SelectAutocomplete> unit tests
